레벨업 경험치 벨런싱

+++ Assets/2DPlatformerRND/01_Scripts/Managers/EnemyStats.cs
@@ -14,7 +14,7 @@ namespace PahlBit
         public float AttackRange;
         public PercentUp ItemDrop;
         public int GoldOnDeath;
-        public float ExpOnDeath;
+        public int ExpOnDeath;
         public Percent HitChance;
 
         public static EnemyStats operator *(EnemyStats stat, SpecOption option)
