+++ Assets/2DPlatformerRND/01_Scripts/UI/InGamePanel.cs
@@ -114,8 +114,8 @@ public class InGamePanel : MonoBehaviour
             mLastMaxMP = maxMP;
         }
 
-        int curExp = (int)(mExperience.CurrentExp - mExperience.ExpAtLevelStart);
-        int maxExp = (int)(mExperience.ExpForNextLevel - mExperience.ExpAtLevelStart);
+        int curExp = mExperience.CurrentExp - mExperience.ExpAtLevelStart;
+        int maxExp = mExperience.ExpForNextLevel - mExperience.ExpAtLevelStart;
         if (mLastCurExp != curExp || mLastMaxExp != maxExp)
         {
             float expRate = (float)curExp / maxExp;
